ext_rad

calculates the exo-atmospheric radiation, in MJ/m^2, for time intervals less than a whole day. It can be used to calculate extraterrestrial radiations for a whole day, but computations are a lot more expensive than using ext_rad_day.

ext_rad(t0, t1, lat, lon, ...)

Arguments

t0, t1

time at the beginning and end of the period.

lat

latitude in degres from -180 to 180 (negative values are west of Greenwich)

lon

longitude in degres from -90 to 90 (negative values are south of Equator)

...

other variables @param j day of the year (1 to 366) (only valid if t0 and t1 are not Date or POSIXct objects)

References

ALLEN, R. G. et al. Crop evapotranspiration - Guidelines for computing crop water requirements - FAO Irrigation and drainage paper 56. FAO: Rome, 1998.

Examples

t0 <- 0:23
t1 <- (t0 + 1) %% 24
ext_rad(t0, t1, -30, -48, j = 90)
#> [1] 0