Calcula a evapotranspiração de referência

Calcula a evapotranspiração de referência (ET0, mm) pelo método de Penman-Monteith padronizado pela FAO (Allen et al., 1998) com ou sem as alterações propostas pela American Society of Civil Engineers (ASCE, 2005)

fao_et0(
  temp,
  rn,
  g,
  gamma,
  u2,
  es,
  e,
  hourly = FALSE,
  mode = "ASCE-ET0",
  day = NA
)

Arguments

temp

temperatura média do ar (ºC) (ver detalhes abaixo)

rn

saldo de radiação [MJ/m²]

g

fluxo de calor no solo [MJ/m²] (ver detalhes)

gamma

coeficiente psicrométrico (ver detalhes)

es

pressão de saturação do vapor de água (kPa) (ver detalhes)

e

pressão parcial do vapor de água (kPa) (ver detalhes)

hourly

FALSE para calcular a evapotranspiração diária (mm/dia) (valor padrão), ou TRUE para calcular a evapotranspiração horária (mm/h)

mode

um dos seguintes valores possíveis 'FAO', 'ASCE-ET0', 'ASCE-ETR' (ver detalhes)

day

válido somente no cálculo da evapotranspiração horária. Um vetor booleano informando para cada registro se é dia ou noite (TRUE para dia e FALSE para noite).

Details

Para fins de padronização, no cálculo para intervalos de tempo de 1 dia ou maiores, deve-se definir temp como a média entre a temperatura máxima e a temperatura mínima do ar medidas no período de tempo em que se deseja calcular a evapotranspiração, e não como a média de todas as medições de temperatura no período (Allen et al., 1998).
Caso não se disponha de medições diretas do parâmetro rn, pode-se estimar o seu valor por meio das funções net_shortwave e net_longwave.
Para o cálculo da evapotranspiração de referência, o albedo da cultura de referência deve ser fixado em 0.23 tanto para a grama quanto para a alfafa (Allen et al., 2005). Caso não se disponha de medições diretas do parâmetro g, Allen et al. (1998) recomendam atribuir o valor de 0 para cálculos de evapotranspiração diária ou 0.1 * rn durante o dia e 0.5 * rn durante a noite, para cálculos de evapotranspiração horária tendo como referência a grama. Caso a referência adotada seja a alfafa e não haja medições diretas do fluxo de calor no solo, deve-se adotar 0.04 * rn durante o dia e 0.2 * rn durante a noite (Allen et al., 2005)
O valor de es pode ser calculado por meio da função tetens_equation. Recomenda-se definir es como a média entre os valores retornados por tetens_equation para a temperatura mínima e para a temperatura máxima no período, para cálculos realizados no intervalo de tempo de 1 dia ou maior.
O parâmetro mode pode ser definido como 'FAO', 'ASCE-ET0' ou 'ASCE-ETR'. O valor padrão é 'ASCE-ET0'. Se definido como 'FAO', o cálculo é realizado com os parâmetros definidos em Allen et al. (1998). Se definido como 'ASCE-ET0', o cálculo considera os parâmetros definidos para a grama de referência conforme Allen et al. (2006). Se definido como 'ASCE-ETR', os parâmetros utilizados correspondem ao da alfafa como cultura de referência.

References

ALLEN, R. G.; PEREIRA, L. S.; RAES, D.; SMITH, M. FAO Irrigation and Drainage Paper no 56. FAO: Rome, 1998.
ALLEN, et al.. A recommendation on standardized surface resistance for hourly calculation of reference ETo by the FAO56 Penman-Monteith method. Agricultural Water Management, 81, p. 1-22, 2006. Available from: <https://www.sciencedirect.com/science/article/abs/pii/S037837740500154X>